优化：将章节ID改为长整型，且不再自增

+++ Sourcecode/Song.Entities/LogForStudentExercise.cs
@@ -20,7 +20,7 @@ namespace Song.Entities {
     		
     		protected Int32 _Cou_ID;
     		
-    		protected Int32 _Ol_ID;
+    		protected Int64 _Ol_ID;
     		
     		protected String _Lse_UID;
     		
@@ -108,7 +108,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Ol_ID {
+    		public Int64 Ol_ID {
     			get {
     				return this._Ol_ID;
     			}
@@ -347,7 +347,7 @@ namespace Song.Entities {
     				this._Cou_ID = reader.GetInt32(_.Cou_ID);
     			}
     			if ((false == reader.IsDBNull(_.Ol_ID))) {
-    				this._Ol_ID = reader.GetInt32(_.Ol_ID);
+    				this._Ol_ID = reader.GetInt64(_.Ol_ID);
     			}
     			if ((false == reader.IsDBNull(_.Lse_UID))) {
     				this._Lse_UID = reader.GetString(_.Lse_UID);
@@ -445,7 +445,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Cou_ID = new WeiSha.Data.Field<LogForStudentExercise>("Cou_ID");
     			
     			/// <summary>
-    			/// 字段名：Ol_ID - 数据类型：Int32
+    			/// 字段名：Ol_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Ol_ID = new WeiSha.Data.Field<LogForStudentExercise>("Ol_ID");
     			
@@ -516,3 +516,4 @@ namespace Song.Entities {
     		}
     	}
     }
+
