+++ Sourcecode/Song.Entities/LogForStudentQuestions.cs
@@ -18,7 +18,7 @@ namespace Song.Entities {
     		
     		protected Int32 _Cou_ID;
     		
-    		protected Int32 _Ol_ID;
+    		protected Int64 _Ol_ID;
     		
     		protected DateTime _Lsq_CrtTime;
     		
@@ -78,7 +78,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Ol_ID {
+    		public Int64 Ol_ID {
     			get {
     				return this._Ol_ID;
     			}
@@ -204,7 +204,7 @@ namespace Song.Entities {
     				this._Cou_ID = reader.GetInt32(_.Cou_ID);
     			}
     			if ((false == reader.IsDBNull(_.Ol_ID))) {
-    				this._Ol_ID = reader.GetInt32(_.Ol_ID);
+    				this._Ol_ID = reader.GetInt64(_.Ol_ID);
     			}
     			if ((false == reader.IsDBNull(_.Lsq_CrtTime))) {
     				this._Lsq_CrtTime = reader.GetDateTime(_.Lsq_CrtTime);
@@ -270,7 +270,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Cou_ID = new WeiSha.Data.Field<LogForStudentQuestions>("Cou_ID");
     			
     			/// <summary>
-    			/// 字段名：Ol_ID - 数据类型：Int32
+    			/// 字段名：Ol_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Ol_ID = new WeiSha.Data.Field<LogForStudentQuestions>("Ol_ID");
