+++ Sourcecode/Song.Entities/LogForStudentStudy.cs
@@ -20,7 +20,7 @@ namespace Song.Entities {
     		
     		protected Int32 _Cou_ID;
     		
-    		protected Int32 _Ol_ID;
+    		protected Int64 _Ol_ID;
     		
     		protected String _Lss_UID;
     		
@@ -102,7 +102,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Ol_ID {
+    		public Int64 Ol_ID {
     			get {
     				return this._Ol_ID;
     			}
@@ -305,7 +305,7 @@ namespace Song.Entities {
     				this._Cou_ID = reader.GetInt32(_.Cou_ID);
     			}
     			if ((false == reader.IsDBNull(_.Ol_ID))) {
-    				this._Ol_ID = reader.GetInt32(_.Ol_ID);
+    				this._Ol_ID = reader.GetInt64(_.Ol_ID);
     			}
     			if ((false == reader.IsDBNull(_.Lss_UID))) {
     				this._Lss_UID = reader.GetString(_.Lss_UID);
@@ -394,7 +394,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Cou_ID = new WeiSha.Data.Field<LogForStudentStudy>("Cou_ID");
     			
     			/// <summary>
-    			/// 字段名：Ol_ID - 数据类型：Int32
+    			/// 字段名：Ol_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Ol_ID = new WeiSha.Data.Field<LogForStudentStudy>("Ol_ID");
