+++ Sourcecode/Song.Entities/TestPaperItem.cs
@@ -24,7 +24,7 @@ namespace Song.Entities {
     		
     		protected String _Org_Name;
     		
-    		protected Int32 _Ol_ID;
+    		protected Int64 _Ol_ID;
     		
     		/// <summary>
     		/// -1
@@ -124,7 +124,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Ol_ID {
+    		public Int64 Ol_ID {
     			get {
     				return this._Ol_ID;
     			}
@@ -217,7 +217,7 @@ namespace Song.Entities {
     				this._Org_Name = reader.GetString(_.Org_Name);
     			}
     			if ((false == reader.IsDBNull(_.Ol_ID))) {
-    				this._Ol_ID = reader.GetInt32(_.Ol_ID);
+    				this._Ol_ID = reader.GetInt64(_.Ol_ID);
     			}
     		}
     		
@@ -286,7 +286,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Org_Name = new WeiSha.Data.Field<TestPaperItem>("Org_Name");
     			
     			/// <summary>
-    			/// 字段名：Ol_ID - 数据类型：Int32
+    			/// 字段名：Ol_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Ol_ID = new WeiSha.Data.Field<TestPaperItem>("Ol_ID");
     		}
