+++ Sourcecode/Song.ServiceInterfaces/IStudent.cs
@@ -181,7 +181,7 @@ namespace Song.ServiceInterfaces
         /// <param name="playTime">播放进度</param>
         /// <param name="studyInterval">学习时间，此为时间间隔，每次提交学习时间加这个数</param>
         /// <param name="totalTime">视频总长度</param>
-        void LogForStudyFresh(int couid, int olid, Accounts st, int playTime, int studyInterval, int totalTime);
+        void LogForStudyFresh(int couid, long olid, Accounts st, int playTime, int studyInterval, int totalTime);
         /// <summary>
         /// 记录学员学习时间
         /// </summary>
@@ -192,14 +192,14 @@ namespace Song.ServiceInterfaces
         /// <param name="studyTime">学习时间，此为累计时间</param>
         /// <param name="totalTime">视频总长度</param>
         /// <returns>学习进度百分比（相对于总时长），如果为-1，则表示失败</returns>
-        double LogForStudyUpdate(int couid, int olid, Accounts st, int playTime, int studyTime, int totalTime);
+        double LogForStudyUpdate(int couid, long olid, Accounts st, int playTime, int studyTime, int totalTime);
         /// <summary>
         /// 根据学员id与登录时生成的Uid返回实体
         /// </summary>
         /// <param name="stid">学员Id</param>
         /// <param name="olid">章节id</param>
         /// <returns></returns>
-        LogForStudentStudy LogForStudySingle(int stid, int olid);
+        LogForStudentStudy LogForStudySingle(int stid, long olid);
         /// <summary>
         /// 返回记录
         /// </summary>
@@ -216,7 +216,7 @@ namespace Song.ServiceInterfaces
         /// <param name="platform">平台，PC或Mobi</param>
         /// <param name="count"></param>
         /// <returns></returns>
-        LogForStudentStudy[] LogForStudyCount(int orgid, int couid, int olid, int stid, string platform, int count);
+        LogForStudentStudy[] LogForStudyCount(int orgid, int couid, long olid, int stid, string platform, int count);
         /// <summary>
         /// 分页获取
         /// </summary>
